Enhancement: Skip more tasks when skipping install
When not installing, /usr/local/bin/blackbox_exporter is not guaranteed to
be present. Which means that applying cap_net_raw capability to it
may fail and should be skipped too. This means that installing the
libcap2-bin package may be unnecessary, if not contradicting the
principle of the 'skip_install' setting. And the same goes for the
installation of the python3-apt package in the preflight script.

All of the above are irrelevant when using this role to setup the
configuration of a container which comes with everything pre-installed.

Signed-off-by: Thomas Venieris <thomas.venieris@gmail.com>