feature: allow downloading of package by remote host
Currently node_exporter install always downloads the package on the
localhost and unpacks it there then uploads it to the remote host.

In the case the ansible controller is at the other end of a thin
network pipe, or if it may not even have access to the package URL,
it may be desirable to download and unpack it on the remote host
instead.

This change adds that ability as an optional behavior to _common/install
and wires it up for use by node_exporter, using the
node_exporter_remote_download variable. The default for node exporter
and the behavior for any other roles remains the same as before:
delegate to localhost.

Fixes #389.

Signed-off-by: Travis Downs <travis.downs@gmail.com>