feat(process_exporter): add support for recheck-with-time-limit
This CLI flag is useful to prevent process_exporter from putting
processes in the wrong groups if it happens to scan them for the first
time before they change their command line. This can happen for
processes which call exec(3) for example.

The default setting is 0, which leaves the feature disabled. Note that
setting a non-zero value for --recheck-with-time-limit automatically
also sets --recheck so there's no need to set the two CLI flags.

The flag is available since process-exporter 0.8.0.

Signed-off-by: François HORTA <fhorta@scaleway.com>